Move EncryptionKeyPairStore one level up.
