+++ src/main/kotlin/org/wfanet/measurement/reporting/service/api/EncryptionKeyPairStore.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.reporting.service.api.v1alpha
+package org.wfanet.measurement.reporting.service.api
 
 import com.google.common.hash.Hashing.goodFastHash
 import com.google.protobuf.ByteString
