+++ src/test/kotlin/org/wfanet/measurement/reporting/service/api/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_test")
+
+kt_jvm_test(
+    name = "EncryptionKeyPairStoreTest",
+    srcs = ["EncryptionKeyPairStoreTest.kt"],
+    associates = [
+        "//src/main/kotlin/org/wfanet/measurement/reporting/service/api:encryption_key_pair_store",
+    ],
+    data = [
+        "//src/main/k8s/testing/secretfiles:secret_files",
+    ],
+    deps = [
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto/testing",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/crypto/tink",
+    ],
+)
