+++ src/test/kotlin/org/wfanet/measurement/reporting/service/api/EncryptionKeyPairStoreTest.kt
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.reporting.service.api.v1alpha
+package org.wfanet.measurement.reporting.service.api
 
-import com.google.common.truth.Truth.assertThat
+import com.google.common.truth.Truth
 import com.google.protobuf.ByteString
 import com.google.protobuf.kotlin.toByteStringUtf8
 import java.nio.file.Path
@@ -68,7 +68,7 @@ class EncryptionKeyPairStoreTest {
   fun `InMemoryEncryptionKeyPairStore returns null for non-existing public key`() {
     val keyPairStore = InMemoryEncryptionKeyPairStore(getKeyPairs())
 
-    assertThat(
+    Truth.assertThat(
         runBlocking { keyPairStore.getPrivateKeyHandle(PRINCIPAL_NAME, NON_EXISTENT_PUBLIC_KEY) }
       )
       .isNull()
@@ -78,7 +78,7 @@ class EncryptionKeyPairStoreTest {
   fun `InMemoryEncryptionKeyPairStore returns null for non-existing principal`() {
     val keyPairStore = InMemoryEncryptionKeyPairStore(getKeyPairs())
 
-    assertThat(
+    Truth.assertThat(
         runBlocking { keyPairStore.getPrivateKeyHandle(NON_EXISTENT_PRINCIPAL_NAME, PUBLIC_KEY_1) }
       )
       .isNull()
@@ -105,6 +105,6 @@ class EncryptionKeyPairStoreTest {
     val publicKeyHandle = TinkPublicKeyHandle(publicKey)
     val encryptedText = publicKeyHandle.hybridEncrypt(PLAIN_TEXT)
     val decryptedText = privateKeyHandle.hybridDecrypt(encryptedText)
-    assertThat(decryptedText).isEqualTo(PLAIN_TEXT)
+    Truth.assertThat(decryptedText).isEqualTo(PLAIN_TEXT)
   }
 }
